refactor(reco-gui): extract export worker + FRICTION entries
Extracted run_export() + ExportOutcome + ExportTelemetrySink into
export.rs (270 lines). main.rs reduced from 3441 to 3163 lines.
The export module is fully self-contained: owns its telemetry sink,
codec/quality parsing, progress callbacks, and autocam wiring.

Added FRICTION entries:
- N16: recording doubles GPU render cost (GUI-specific, CLI is smooth)
- N17: no ROI visualization or editing in GUI
- N18: Slint slider loses pointer tracking inside ScrollView
